Incredible India Java Project
+# Incredible_India_Project
This  project is build using java and applets. It involves  an applet that displays famous places of states of India. This applet includes buttons for each state that, when clicked, display a picture of a famous place in that state. This applet also include a text area that displays a brief description of the place when the button is clicked.